requirements.txt
streamlit>=1.33
pandas>=2.0
openpyxl>=3.1
psycopg2-binary>=2.9